+++ config/index.conf.php
@@ -14,7 +14,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
- * Copyright (c) 2018 (original work) Open Assessment Technologies SA;
+ * Copyright (c) 2020 (original work) Open Assessment Technologies SA;
  */
 
 declare(strict_types=1);
@@ -86,8 +86,7 @@ return [
                 ],
                 'properties' => [
                     'class' => [
-                        'type' => 'keyword',
-                        'ignore_above' => 256,
+                        'type' => 'text',
                     ],
                     'content' => [
                         'type' => 'text',
@@ -136,8 +135,7 @@ return [
                 ],
                 'properties' => [
                     'class' => [
-                        'type' => 'keyword',
-                        'ignore_above' => 256,
+                        'type' => 'text',
                     ],
                     'label' => [
                         'type' => 'text',
@@ -179,8 +177,7 @@ return [
                 ],
                 'properties' => [
                     'class' => [
-                        'type' => 'keyword',
-                        'ignore_above' => 256,
+                        'type' => 'text',
                     ],
                     'label' => [
                         'type' => 'text',
@@ -222,8 +219,7 @@ return [
                 ],
                 'properties' => [
                     'class' => [
-                        'type' => 'keyword',
-                        'ignore_above' => 256,
+                        'type' => 'text',
                     ],
                     'label' => [
                         'type' => 'text',
@@ -265,8 +261,7 @@ return [
                 ],
                 'properties' => [
                     'class' => [
-                        'type' => 'keyword',
-                        'ignore_above' => 256,
+                        'type' => 'text',
                     ],
                     'label' => [
                         'type' => 'text',