+++ CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 -----
 - README.md updated with indexing script
 - Class properties automatically indexed
+- Creted separated indexes for Items, Tests, Test-Takers, Groups and Deliveries
+- Added a query builder to handle searches on different indexes
 
 1.1.1
 -----