+++ CHANGELOG.md
@@ -3,6 +3,8 @@ CHANGELOG
 
 1.2.0
 -----
+- README.md updated with indexing script
+- Class properties automatically indexed
 - Creted separated indexes for Items, Tests, Test-Takers, Groups and Deliveries
 - Added a query builder to handle searches on different indexes